feat: Implement insertions
This commit implements the database insertions. Each table is populated
with the data gathered from the amazon-meta.txt file. We try to make
this process a little bit faster by using threads to parallelize the
tabe insertion, unfortunately is not possible to do it for all table as
some tables depend on data in another table (foreign key).

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>